Add some more fixtures; split into multiple modules
