+++ test/fixtures/Droid.ts
@@ -1,0 +1,12 @@
+export interface Droid {
+  name:string;
+  primaryFunction:Droid.Function;
+}
+
+
+export namespace Droid {
+  export enum Function {
+    PROTOCOL,
+    MILITARY
+  }
+}
