+++ test/fixtures/common.ts
@@ -1,0 +1,6 @@
+export enum Seasons {
+  SPRING   = "SPRING",
+  SUMMER   = "SUMMER",
+  FALL     = "FALL",
+  WINTER   = "WINTER",
+}
